red4ext_rs
Module log

Convenience logging macros. By default all macros require a SdkEnv instance to be passed as the first argument. If the log feature is enabled, this module becomes an alias to the log crate.

Macros

debug

Logs a message at the debug level.

error

Logs a message at the error level.

info

Logs a message at the info level.

trace

Logs a message at the trace level.

warn

Logs a message at the warn level.