red4ext_rs
Trait IntoRepr

pub trait IntoRepr: Sized {
    type Repr: NativeRepr;

    // Required method
    fn into_repr(self) -> Self::Repr;
}

A trait for types that can be converted into a representation that can be passed across the FFI boundary to the game.

Required Associated Types

type Repr: NativeRepr

Required Methods

fn into_repr(self) -> Self::Repr

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.