Struct red4ext_rs::types::Ref

pub struct Ref<T: ScriptClass>(/* private fields */);

A strong reference to a script class. A live instance of the reference guarantees that the instance is still alive. When the reference is dropped, the strong reference count is decremented. If the strong reference count reaches zero, the underlying value is destroyed.

Implementations

impl<T: ScriptClass> Ref<T>

pub fn new() -> Option<Self>

Creates a new reference to the class.

pub fn new_with(init: impl FnOnce(&mut T)) -> Option<Self>

Creates a new reference to the class and initializes it with the provided function.

pub unsafe fn fields(&self) -> Option<&T>

Returns a reference to the fields of the class.

Safety

The underlying value can be accessed mutably at any point through another copy of the Ref. Ideally, the caller should ensure that the returned reference is short-lived.

pub unsafe fn fields_mut(&mut self) -> Option<&mut T>

Returns a mutable reference to the fields of the class.

Safety

The underlying value can be accessed mutably at any point through another copy of the Ref. Ideally, the caller should ensure that the returned reference is short-lived.

pub unsafe fn instance( &self ) -> Option<&<<T as ScriptClass>::Kind as ClassKind<T>>::NativeType>

Returns a reference to the instance of the class.

Safety

The underlying value can be accessed mutably at any point through another copy of the Ref. Ideally, the caller should ensure that the returned reference is short-lived.

pub fn downgrade(self) -> WeakRef<T>

Converts the reference to a WeakRef. This will decrement the strong reference count and increment the weak reference count.

pub fn cast<U>(self) -> Option<Ref<U>>
where U: ScriptClass,

Attempts to cast the reference to a reference of another class. Returns None if the target class is not compatible.