Trait red4ext_rs::NativeRepr

pub unsafe trait NativeRepr {
    const NAME: &'static str;
}

A trait for types that can be passed across the FFI boundary to the game engine without any conversion.

Safety

Implementations of this trait are only valid if the memory representation of Self is idetical to the representation of type with name Self::NAME in-game.

Required Associated Constants

const NAME: &'static str

Object Safety

This trait is not object safe.