red4ext_rs
Trait Plugin

pub trait Plugin {
    const NAME: &'static U16CStr;
    const AUTHOR: &'static U16CStr;
    const VERSION: SemVer;
    const SDK: SdkVersion = SdkVersion::LATEST;
    const RUNTIME: RuntimeVersion = RuntimeVersion::RUNTIME_INDEPENDENT;
    const API_VERSION: ApiVersion = ApiVersion::LATEST;

    // Provided methods
    fn exports() -> impl Exportable { ... }
    fn on_init(_env: &SdkEnv) { ... }
}

A definition of a RED4ext plugin.

Required Associated Constants

const NAME: &'static U16CStr

The name of the plugin.

const AUTHOR: &'static U16CStr

The author of the plugin.

const VERSION: SemVer

The version of the plugin.

Provided Associated Constants

const SDK: SdkVersion = SdkVersion::LATEST

The RED4ext SDK version the plugin was built with.

const RUNTIME: RuntimeVersion = RuntimeVersion::RUNTIME_INDEPENDENT

The version of the game the plugin is compatible with.

const API_VERSION: ApiVersion = ApiVersion::LATEST

The RED4ext API version.

Provided Methods

fn exports() -> impl Exportable

A list of definitions to be exported automatically when the plugin is loaded. This can be used to define classes and functions that will available to use in the game. See the exports! macro for more information.

fn on_init(_env: &SdkEnv)

A function that is called when the plugin is initialized.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.