red4ext_rs

Module types

A module encapsulating various types defined in the RED4ext SDK.

Re-exports

pub use array::RedArray;

Modules

array

A dynamically sized array.

Structs

ArrayType

An RTTI representation of an array type.

Bitfield

An RTTI representation of a bitfield type.

CName

A hash representing an immutable string stored in a global name pool.

CNamePool

A global pool containing all CNames.

Class

An RTTI representation of a class.

ClassFlags

Flags for a class type.

ClassHandle

Class handle to be used to register a class with RttiSystemMut.

Cruid

Curve

CurveType

An RTTI representation of a curve type.

DataBuffer

DateTime

DeferredDataBuffer

EditorObjectId

EngineTime

EntityId

Enum

An RTTI representation of an enum type.

Function

An RTTI representation of a function type.

FunctionFlags

Flags for a function.

GameEngine

GameInstance

Scripted game instance.

GameTime

GlobalFunction

An RTTI representation of a global function.

Guid

IAllocator

An interface for allocating and freeing memory.

IScriptable

The base type for all scripted classes.

ISerializable

The base type for all types held inside game references.

InvokeStatic

InvokeVirtual

ItemId

LocalizationString

MessageResourcePath

Method

An RTTI representation of a method.

MultiChannelCurve

NativeArrayType

An RTTI representation of a native array type.

NativeClass

An RTTI representation of a native class.

NativeGameInstance

Native game instance.

NodeRef

PointerType

An RTTI representation of a pointer type.

PoolRef

A reference to a value stored in a pool.

Property

An RTTI representation of a property.

PropertyFlags

Flags for a property.

RaRef

An asynchronous resource reference.

RaRefType

An RTTI representation of an asynchronous resource reference type.

RedHashMap

A hash map.

RedString

A dynamically allocated string.

Ref

A reference counted shared pointer to a script class.

RefType

An RTTI representation of a reference type.

ResRef

ResourceRef

ResourceRefType

An RTTI representation of a resource reference type.

RwSpinLockReadGuard

A read-write spin lock read guard. Permits any number of readers to access the locked data.

RwSpinLockWriteGuard

A read-write spin lock write guard. Permits only one thread at a time to access the locked data.

ScriptRef

A reference to local script data.

ScriptRefType

An RTTI representation of a script reference type.

ScriptableSystem

SharedDataBuffer

StackArg

A stack argument to be passed to a function.

StackFrame

A script stack frame.

StaticArray

StaticArrayType

An RTTI representation of a static array type.

StaticMethod

An RTTI representation of a static method.

TweakDbId

Type

An RTTI representation of a type.

ValueContainer

A container holding values, e.g. object fields or function arguments.

ValuePtr

A pointer to a script value.

Variant

WeakRef

A weak reference to a script class. Before use, it must be upgraded to a strong reference using WeakRef::upgrade.

WeakRefType

An RTTI representation of a weak reference type.

Enums

GameEItemIdFlag

see gameEItemIDFlag and CET initialization.

GamedataItemStructure

Opt

A convenience type to explicitly mark a function argument as opt T.

TaggedType

Enum representation of the RTTI type.

TypeKind

An identifier for a kind of type.

Constants

CALL_INSTR_SIZE

OPCODE_SIZE

Traits

Hash

A trait for types that can be hashed.

Instr

A trait for types that correspond to bytecode instructions.

Poolable

A trait for types that can be stored in a pool.

PoolableOps

A trait with operations for types that can be stored in a pool.

Type Aliases

FunctionHandler

A handler for function calls.

OpcodeHandler

A function pointer type for bytecode opcode handlers.