Module red4ext_rs::types

A module encapsulating various types defined in the RED4ext SDK.

Structs

ArrayType

Bitfield

CName

A hash representing an immutable string stored in a global name pool.

CNamePool

Class

ClassFlags

ClassHandle

Class handle to be used to register a class with RttiSystemMut.

Cruid

Curve

CurveType

DataBuffer

DateTime

DeferredDataBuffer

EditorObjectId

EngineTime

EntityId

Enum

Function

FunctionFlags

GameTime

GlobalFunction

Guid

IAllocator

An interface for allocating and freeing memory.

IScriptable

ISerializable

IntoIter

InvokeStatic

InvokeVirtual

ItemId

LocalizationString

MessageResourcePath

Method

MultiChannelCurve

Native

A marker type for native classes. Native classes are represented directly by their native type.

NativeArrayType

NativeClass

NodeRef

PointerType

PoolRef

A reference to a value stored in a pool.

Property

PropertyFlags

RaRef

RaRefType

RedArray

A dynamically sized array.

RedHashMap

RedString

A dynamically allocated string.

Ref

A strong reference to a script class. A live instance of the reference guarantees that the instance is still alive. When the reference is dropped, the strong reference count is decremented. If the strong reference count reaches zero, the underlying value is destroyed.

RefType

ResRef

ResourceRef

ResourceRefType

RwSpinLockReadGuard

RwSpinLockWriteGuard

ScriptRef

A reference to local script data.

ScriptRefType

Scripted

A marker type for scripted classes. Scripted classes are stored with an additional level of indirection inside of IScriptable.

SharedDataBuffer

StackArg

A stack argument to be passed to a function.

StackFrame

A script stack frame.

StaticArray

StaticArrayType

StaticMethod

TweakDbId

Type

ValueContainer

ValuePtr

Variant

WeakRef

A weak reference to a script class. Before use, it must be upgraded to a strong reference using WeakRef::upgrade.

WeakRefType

Enums

GameEItemIdFlag

see gameEItemIDFlag and CET initialization.

GamedataItemStructure

Kind

TaggedType

Constants

CALL_INSTR_SIZE

OPCODE_SIZE

Traits

ClassKind

A trait for distinguishing between native and scripted classes.

Hash

Instr

A trait for types that correspond to bytecode instructions.

Poolable

A trait for types that can be stored in a pool.

PoolableOps

A trait with operations for types that can be stored in a pool.

ScriptClass

A trait for types that represent script classes.

ScriptClassOps

A trait for operations on script classes.

Type Aliases

FunctionHandler

OpcodeHandler

A function pointer type for bytecode opcode handlers.