Struct red4ext_rs::types::Class

pub struct Class(/* private fields */);

An RTTI representation of a class.

Implementations

impl Class

pub fn name(&self) -> CName

pub fn flags(&self) -> ClassFlags

pub fn set_flags(&mut self, flags: ClassFlags)

pub fn size(&self) -> u32

pub fn holder_size(&self) -> u32

pub fn alignment(&self) -> u32

pub fn properties_size(&self) -> u32

pub fn properties(&self) -> &RedArray<&Property>

pub fn cached_properties(&self) -> &RedArray<&Property>

pub fn methods(&self) -> &RedArray<&Method>

pub fn method_map(&self) -> &RedHashMap<CName, &Method>

pub fn static_methods(&self) -> &RedArray<&StaticMethod>

pub fn get_method( &self, name: CName, ) -> Result<&Method, impl Iterator<Item = &Method>>

Resolves a method by name. Returns the method wrapped in Ok if it exists in this class or any of its base classes. If the method is not found, this function returns an iterator over all methods with a matching short name wrapped in Err.

pub fn base(&self) -> Option<&Class>

pub fn base_iter(&self) -> impl Iterator<Item = &Class>

pub fn all_properties(&self) -> impl Iterator<Item = &Property>

pub fn is_class(&self) -> bool

pub fn instantiate(&self) -> ValueContainer

pub fn add_method(&mut self, func: PoolRef<Method>)

pub fn add_static_method(&mut self, func: PoolRef<StaticMethod>)

pub fn add_property(&mut self, prop: PoolRef<Property>)

pub fn as_type(&self) -> &Type

pub fn as_type_mut(&mut self) -> &mut Type